move DOM test to own file
