+++ src/platform/user/tests/authentication/components/loginButtons.unit.spec.js
@@ -1,0 +1,47 @@
+import React from 'react';
+import { expect } from 'chai';
+import { mount } from 'enzyme';
+import sinon from 'sinon';
+
+import * as authUtilities from '../../../authentication/utilities';
+import NewDesignButtons from '../../../authentication/components/NewDesignButtons';
+import { CSP_IDS } from '../../../authentication/constants';
+
+describe('login DOM ', () => {
+  const sandbox = sinon.createSandbox();
+
+  beforeEach(function() {
+    sandbox.spy(authUtilities, 'login');
+  });
+
+  afterEach(() => {
+    sandbox.restore();
+  });
+
+  it('login buttons should properly call login method', () => {
+    const loginButtons = mount(<NewDesignButtons />);
+
+    const testButton = button => {
+      const loginCSP = button.prop('data-csp');
+
+      const expectedArgs = {
+        idme: { policy: CSP_IDS.ID_ME },
+        logingov: { policy: CSP_IDS.LOGIN_GOV },
+        dslogon: { policy: CSP_IDS.DS_LOGON },
+        mhv: { policy: CSP_IDS.MHV },
+      };
+
+      button.simulate('click');
+
+      expect(authUtilities.login.calledOnce).to.be.true;
+      expect(authUtilities.login.getCall(0).args[0]).to.deep.equal(
+        expectedArgs[loginCSP],
+      );
+
+      sandbox.reset();
+    };
+
+    loginButtons.find('button').forEach(testButton);
+    loginButtons.unmount();
+  });
+});
